requirements.txt
validators
whitenoise==6.0.0
six==1.14.0
pytz==2021.3
python-dateutil==2.8.2
phonenumberslite==8.12.40
Pillow==9.0.0
mysql-connector-python<=8.0.26
future==0.16.0
gunicorn==20.1.0
Django==3.2.9
django-allauth==0.46.0
django-environ==0.8.1
django-jsonfield==1.4.1
django-phonenumber-field==6.0.0
django-redis==5.2.0
django-timezone-field==4.2.3
Babel==2.9.1
requests
django-cleanup
xhtml2pdf
django-mathfilters
django-online-users
django-ckeditor
django-debug_toolbar
channels
asgiref==3.5.2
redis== 4.3.4
msgpack==1.0.4
asgi_redis
timeago
daphne==3.0.2
psycopg2
dj-database-url
djangorestframework
channels-redis
django-clearcache
PyPDF2